hsm: Prevent stop monitoring of a pool domain
When hosted engine agent is entering local maintenance, it must stop
monitoring the hosted engine domain. This is racy because the domain may
be in the middle of an import operation.

To avoid this race, stopMonitoringDomain will fail now with new
StorageDomainIsMemberOfPool exception, so hosted engine agent can stop
monitoring safely.

This failure is also important even without this race; if a domain is
owned by the pool, nobody should be able to stop monitoring it. The only
way to stop monitoring a pool domain is disconnecting from the pool.

Change-Id: I99555b230dfdab41384af0461e4b30c590b669b9
Bug-Url: https://bugzilla.redhat.com/1282187
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/49387
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>