lib: daemon: autodetect online cpus for affinity
In commit a1d4e23 we enabled by default the cpu affinity
for VDSM, after we had confirmation this delivered important
performance benefits.

An hardcoded default setting for any cpu core is a too naive
approach, as it could break automated setups if the said CPU
is offline. And Vdsm cannot guess that.

To solve this, we add a function to learn from the kernel which
cpus are online, using a file from /sys pseudo-filesystem.
This is a simplified version of the same approach libvirt uses.
We don't leverage libvirt because it is too early in the
VDSM startup.

The layout of /sys, like /proc, is part of the kernel ABI,
so there are only small breakage concerns [1][2].
Furthermore, we already rely on /sys for other VDSM flows,
so this patch doesn't add fragility

We add a special value to the cpu_affinity tunable, to let
Vdsm select automatically best cpu. We choose the second
online core in the system, handling the case on which
only one core is active.

The admin can still select explicitely the cores on which
Vdsm should pin to, or still can disable the pinning entirely.

[1] https://www.kernel.org/doc/Documentation/ABI/README
[2] https://www.kernel.org/doc/Documentation/ABI/testing/sysfs-devices-system-cpu

Change-Id: Iba834f67a43ce766308036cbd079107340ed69d8
Bug-Url: https://bugzilla.redhat.com/1286462
Bug-Url: https://bugzilla.redhat.com/1279431
Backport-To: 3.6
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/49402
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/49612