spec: Require newer libvirt version for el7
Libvirt 1.2.17 introduced a regression where physical volume size is
always zero. This breaks our extension logic, using:

    physical - alloc < drive.watermarkLimit

As a result, we try to extend the drive on each check to next size after
0 (1GiB), spamming the logs with lvm warnings. Finally the vm is paused,
since we never extend the disk.

Libvirt 1.2.17-5 fixed this issue. We require now this version.

We don't require the package yet on CentOS, since it is not available
yet.

The issue is probably effecting also Fedora 23. We will provide a
separate patch for Fedora if needed.

This patch is also squashed with change
I0e6096df3be73b6929940460c9467d9bb2ab1768 in order to avoid the missing
package error introduced in the original patch

Change-Id: Iea42288e779ba8a63968087d1df27e0f3665e406
Backport-To: 3.6
Bug-Url: https://bugzilla.redhat.com/1251008
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44834
Continuous-Integration: Jenkins CI
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Signed-off-by: Tal Nisan <tnisan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/49444
Reviewed-by: Francesco Romani <fromani@redhat.com>