daemon: revert cpu-affinity enabling by default
In commit a1d4e23 we enabled by default the cpu affinity
for VDSM, after we had confirmation this delivered important
performance benefits.

Unfortunately, the default setting, was too naive
for large, complex setup, like the ones we first seen on PPC64
hosts. VDSM tries to pin itself to cpu #1, but if this cpu
is offline, it fails to start.

This was indeed expected behaviour, and the fix is trivial (just
reconfigure /etc/vdsm/vdsm.conf), but it is very unpractical in large
settings, where the admin has to manually fix many configuration
files. This obviously breaks automated flow.

Unfortunately, is too late in the release schedule to add smarter
CPU selection logic (e.g. patch 49402), so to avoid VDSM start
failures, we revert the default, to fall back to a safe state.

Change-Id: I91e3753172b4fa4d60686b9cf21e596d2f408393
Related-To: https://gerrit.ovirt.org/#/c/49402
Related-To: https://bugzilla.redhat.com/1279431
Backport-To: 3.6
Bug-Url: https://bugzilla.redhat.com/1286462
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/49463
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI