vm: enabling memory snapshot without disks
Taking a memory snapshot of a VM without disks should be
supported (or merely with Cinder/LUN disks for that matter).

Hence, removing the early return on 'VM -> snapshot'
in case 'newDrives' is empty.

Instead, proper logging will be added on consecutive patches.

Possible configurations and settings that lead to this flow:
* Unchecking all the disks in the create snapshot dialog.
* VM with only cinder disks.
* VM without disks.

Change-Id: Ia540fa9009f627f7a2943ef393084eee3f047bf5
Bug-Url: https://bugzilla.redhat.com/1287066
Signed-off-by: Daniel Erez <derez@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/50199
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>