net: move qos reporting to netinfo
Up until now, qos reporting was part of caps.py, which lies in an
upper layer then netinfo, just because the tc code could not be used
from within netinfo. Now it is possible to bury this reporting code
inside netinfo new package, where it belongs.
Note, that this change creates a circular import because the
configurators/qos.py code is used from netinfo/qos.py but is trying to
import netinfo itself. This was solved by moving _root_qdisc to
netinfo/qos.py. This actually makes sense, because nobody should be
dependant on network/configurators for the sake of reporting stuff-
especially not the main reporting code (netinfo package). The only qos
reporting code is now left in vdsm/tc.

Change-Id: Ie3526c4ae609f29561904bddacc1e9f9194435e9
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/49620
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI