net: Compare kernel vs running bond config opt vals
Comparing between the bond running config and the
bond kernel config should consider that some values
may be inputted as name or numeric.
A dictionary is kept to translate between the name
and its numeric value.

The running config bonding option values are normalized
to their numeric values and compared to the kernel
config which represents the values as numeric.

Change-Id: I0f334a7ae2ecd8fc75c05972e37d539cf7c28661
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/49579
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>