getAllImages: Use manifest attributes when constructing glob pattern
getAllImages searches a domain's images directory for directories that
look like a UUID.  To make this function testable with domains that are
not located at the usual storage repo location '/rhev/data-center', find
the images directory using the StorageDomainManifest's mountpoint
attribute.  Note that this is how getAllVolumes already works.

Change-Id: I8e69d3096a1b4c20dcee33e96c8a7bdd1b8dcaa9
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/49684
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI