testlib: Allow recording class methods
The @recorded decorator could not be applied to class methods, and we
are using lot of boilerplate code to record class methods.

This patch modify the decorator so it can be used to decorate both
instance and class methods.

Since class attributes are visible via the instance, we use
__recording__ for the instance, and __class_recording__ for the class.

Change-Id: Idf33cba4ab41dc4cc7cc09d4afa70d33b75cde9f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/49960
Continuous-Integration: Jenkins CI
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>