vdsm: introduce cpuinfo module
The objective of the new modules is encapsulating functionality related
to CPU from bare-metal standpoint. That means parsing cpuinfo and
exposing host architecture. This approach cleans up 'caps' module and,
more importantly, moves the code to vdsm library - allowing hooks to
work with cpu info and architecture. This allows us to move fake_kvm
to a hook.

Change-Id: Iaa702b05f3825ebdcfed16d86d39a8c38fcf224c
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46912
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>