caps/lib: move CPU architecture details to lib/cpuarch
CPU architecture details include things such as
* current CPU architecture,
* the CPU architecture of VMs and
* supported architectures.

Having this information and functionality inside caps.py means that it
cannot be used in hooks or where circular imports may be a problem. To
solve this, we create new cpuarch module inside sitelib where we move
and slightly clean-up the logic. This way, arch information is
accessible to any VDSM code.

Change-Id: I46bb51769562fbb9d816d190c2aebfd0660cc3b0
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/49972
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>