gluster: Warn instead of failing for unsupported gluster replica modes
A previous patch enforced the use of only supported gluster replication
modes.  If an unsupported mode was requested, the connection attempt
would fail.  Unfortunately, this was a bit heavy-handed and has resulted
in breaking working environments during upgrade. To lessen the impact on
users, instead of failing, log a warning. A related patch to
ovirt-engine will warn users about the unsupported replication mode in
the UI.

Change-Id: I359e2405153f45b6ae303e00eb0e04a5ae14cb99
Signed-off-by: Ala Hino <ahino@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1286565
Reviewed-on: https://gerrit.ovirt.org/50119
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Sahina Bose <sabose@redhat.com>
Continuous-Integration: Jenkins CI