virt: sampling: reduce host averaging window
Host samples are stored into a SampleWindow of
size 5 for mostly historical reasons, with
no clear rationale.

However, a clear adverse effect of this setting
is new network devices are reported with a larger
delay, especially since commit 6547ea8b90fa191257
which enlarged the sample interval to 15s from
the previous, too frequent, interval of 2s.

To be reported, a network device must be present in
the first and the last sample.
For one hotplugged device, the delay can be high as
sample_interval * window_size.

To counter this delay, we reduce the sample window
size to two items, which is the minimum meaningful
value.

Change-Id: Ib5f416615058458321a03d7e149b1db243c34002
Bug-Url: https://bugzilla.redhat.com/1261457
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/50148
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>