executor: Fix the task-callable mismtach
Executor is executing queue tasks, but the public api is dealing with
callable and timeout. These leak into the executor internals, and cause
confusion when we log the executor state.

Now we pack the callable and the timeout in a Task namedtuple, fixing
the mismatch, and providing more information when logging the task in
Worker.__repr__.

    Task(callable=<function func at 0x7fd84f3ec500>, timeout=7)

Change-Id: Id50a4175bc9cb27b9e432a2c871b1ba98cf625c8
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/50228
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>