build: Add dependencies for generated .py files
In the past we checked .py.in files with pep8 and pyflakes. Recently we
removed this wrong check, and we are checking only .py files.

This change required running "make; make check", to have .py files
generated before running make check.

This patch adds dependencies so "make check" will generated the files it
wants to check.

Searching for .py.in file is not needed now and was removed.

Change-Id: I57eeba76a5e07a1edeae6869e103547347e874cf
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/50299
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>