jobs: Add error property
When a Job encounters an error, there should be a standardized way for
that error to be stored so that code working with Job objects has a
standard way to propagate the error.

Add an 'error' property to the job class which is None by default but
can contain an instance of utils.GeneralException or one of its
decendants.  If error is set, the dictionary returned by the call to
info() will contain an 'error' key.  The value of this key is a status
code (containing a standard vdsm code and message).

Change-Id: I3f6e27beeeecd69db1673e5168524b0ce94f2e9e
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/50354
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Shahar Havivi <shavivi@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Continuous-Integration: Jenkins CI