jobs: Only report progress if set
Not all Jobs have deterministic progress and some jobs will complete so
quickly that progress updates are not practical.  If progress is not
available do not report it.  Consumers of the Job info could choose to
render a non-deterministic progress indication (barber pole, or spinning
indicator) instead of a regular progress bar.

We expect all jobs to start without any progress to report since they
may be pending (waiting for a worker thread) or initializing (taking
locks).  Jobs that can provide progress information will start to report
later.  Jobs that do not support progress need not implement anything as
the default behavior of Job is to never report progress.

Change-Id: I3cc8b66d825676045fbcb8431bb3cf11885ff02b
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/50355
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Tested-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Shahar Havivi <shavivi@redhat.com>