gluster: Handle missing glusterfs-cli package
In 3.6 we introduced a soft dependency to glusterfs-cli. This was
required in order to get gluster volume info needed for to purposes:
1. get gluster backup servers in order to automatically mount gluster
   back servers
2. get replica related to verify that it is supported
However, during upgrade, if glusterfs-cli is missing, mount failed. In
this case, we'd like to add a warning message clearly indicating about
the missing glustefs-cli package, and ask the admin to install it in
order to enable automatica mount of gluster backup servers.

Change-Id: I9c33aa33e4ffe6a382d40e1bc63f6735efcfcd1f
Signed-off-by: Ala Hino <ahino@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1290997
Reviewed-on: https://gerrit.ovirt.org/50363
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Sahina Bose <sabose@redhat.com>