net: Correctly apply MTU values on networks
Two issues have been resolved by this change:

- New networks with no MTU specification are being
set by default with their connected device (bond,
vlan, nic) mtu (which does not have to be 1500).
Fixed by detecting when no MTU is specified in the
configuration, and adding the default (1500)
explicitly.
The assumption of a single default mtu when one is
not specified in the setup is wrong, causing in
some cases an unnecessary restoration of networks
during network restoration.
- Test fix: The NIC/s mtu should be set to the
maximum mtu of the remaining networks.

Change-Id: Iba24363f5b9ea70392a68885cf5374800ab58549
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/50397
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>