hostdev: fix detection for ppc64le
x86_64 detects the state of hostdev support by looking at 2 places:

* /sys/class/iommu to determine the status of DMA remapping,
* /sys/kernel/iommu_groups to determine that at least some IOMMU groups
  exist.

The issue in ppc64le is related to the first place - /sys/class/iommu
is newer interface and may not be supported on all kernels. 7.2 on
x86_64 seems to be updated enough that we can keep the detection, but
for ppc64le we cannot use it.

Change-Id: I314c8bda6a6831bf334e08085e8bf2fc4874aced
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/50525
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>