storage: add getHostId to DomainMonitor
Some SDM verbs need to acquire the domainLock during their job.  In
order to do this they must be passed the domain's hostId property.  An
easy way to expose this is to add a getHostId method to the
DomainMonitor.

Another method (not chosen) is for IRS to access the host id via the
StorageDomainManifest.getPools method:

    dom_manifest.getPools()[0].id

The main reason for using the domainMonitor is to avoid adding any new
users of the StoragePool class.  Since the domainMonitor acquires the
hostID via SANLock, it seems qualified to communicate that hostID to its
owner.

Change-Id: I21c3c59678ad6a7d6087d05c7109ec4087cbb09c
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/50630
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>