fileSD: Do not try to remove non-existent lease file
The volume.lease file exists only if a domain has volume leases. Avoid
removing the non-existent lease file and the bogus exception in the log
when the operation "fails".

This is the same patch from master, applied to FileStorageDomain instead
of FileStorageDomainManifest, since deleteImage was not moved to the
manifest class in 3.6.

Change-Id: I525704678707dfef40e34497bc4917dd19d73034
Bug-Url: https://bugzilla.redhat.com/1291207
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/50589
Continuous-Integration: Jenkins CI
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Liron Aravot <laravot@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51132
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Idan Shaby <ishaby@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>