fileSD: improve error handling in deleteImage
In FileStorageDomainManifest.deleteImage, we used to log that the volume
could not be removed even if it was removed but the meta/lease file
could not be removed afterwards.

This patch logs more informative errors about the specific file that
could not be removed. If the removal fails because the file doesn't
exist, we log a warning. Otherwise, we log an error message that the
file can't be removed.

Also, if one of the files couldn't be removed, we didn't try to remove
the others.  Now we handle each removal separately, so that a failure in
the removal of one file doesn't affect the removal of the others.

This patch is a backport for patch
I9eeb28a70f708a4f9a5effe4ff294da63b757369 which was reverted.

Change-Id: I22ceee059cce1ec65cd455ce567cba0e49c00db2
Bug-Url: https://bugzilla.redhat.com/1292092
Signed-off-by: Idan Shaby <ishaby@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51229
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51299
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>