gluster: ensure LV size is multiple of VG's PE value
 Alignment value for JBOD is changed to 256 KiB from 0
in https://gerrit.ovirt.org/#/c/47959 and alignment value is
used as PE value in VG creation. It resulted in following
error for some JBOD disks.

"DeviceError: ('Adding new lv is too large to fit in free space',
 u'vg-brick2-pool-brick2'"

Size of LV should be multiple of PE value. But this calculation
was done only for RAID volume not for JBOD and that was the casue for
above error.

With this patch, lv size will be adjusted to the multiples of PE
always regardless of whether its RAID or JBOD.

Change-Id: I3a1609544db36b16b5b8fc0a52d5ef04e770c18b
Bug-Url: https://bugzilla.redhat.com/1270792
Signed-off-by: Ramesh Nachimuthu <rnachimu@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/50802
Reviewed-by: Sahina Bose <sabose@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/50854
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Continuous-Integration: Yaniv Bronhaim <ybronhei@redhat.com>