cpuinfo: add predicates for x86 and ppc platform detection
Previously we used string (to check for x86_64) and tuple (for ppc64 and
ppc64le). This patch adds two predicates to the Architecture class of
caps module:

* is_x86(arch)
* is_ppc(arch)

These classmethods can be used to determine in which class does given
'arch' argument belong - allowing us to query architecture in
consistent and unified way.

Change-Id: I0981367436ddfbfabfd3484019b96831e31f9609
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/50894
Continuous-Integration: Jenkins CI
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>