build: update libvirt dependency for fedora
Require a libvirt version that fixes the problem of reporting
physical=0 for COW2 volumes on block storage on Fedora. The given
version 1.1.2.18.1 exists in Fedora 23 and fixes the issue.

See Fedora BZ: https://bugzilla.redhat.com/1253754

Though the fix is only in libvirt-daemon, the libvirt-python
required version is also aligned to it to keep consistency.

Change-Id: I7569dd69d7c5ccfe7cb275acba64c2506d0a9e0a
Bug-Url: https://bugzilla.redhat.com/1253756
Signed-off-by: Fred Rolland <frolland@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51295
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>