vm: safer removal of the recovery file
If a user shutdowns a VM from inside the guest, Vdsm leaves
a stale recovery file.
The cause is a race between onQemuDeath, being asynchronous, and
Vm.destroy().
The actual culprit is the ordering of events between onQemuDeath and
destroy(), and the shutdown from within the guest is just the easiest
way to trigger this issue.

In the QEMU death path, Vdsm calls releaseVm, which wipes the recovery
file, and after a succesful completion does saveState, which recreates
the recovery file.
Later, on destroy(), Vdsm tries to do releaseVm again, but releaseVm
will not run, because of the flag that the previous succesful
call set. releaseVm must run exactly once.

Correct sequence:
- destroy()
-- vm._destroyed = True
-- releaseVm() - done (vm._released == False)
-- vm._released = True
[later]
- onQemuDeath
-- releaseVm() - skipped (vm._released == True)
-- saveState() - skipped (vm._destroyed == True)

Faulty sequence:
- onQemuDeath()
-- releaseVm() - done (vm._released == False)
-- vm._released = True
-- saveState() - done <BUG!> (vm._destroyed == False)
[later]
- destroy()
-- vm._destroyed = True
-- releaseVm() - skipped (vm._released == True)

The fix is to protect the recoveryFile handling with a lock.

Change-Id: I9a8d2b34599da17bee35c9bb694122ecd9d58068
Backport-To: 3.6
Backport-To: 3.5
Bug-Url: https://bugzilla.redhat.com/1253043
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51387
Continuous-Integration: Jenkins CI
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/52515