testEnablePromisc: wait for promisc event
We have seen cases where testEnablePromisc failed on CI. This could have
been caused if the test is faster than the kernel, and tests the promisc
flag before it was applied to the link.

There might be many more similar races in our tests; this patch should
serve only as a proof of concept.

Change-Id: I8ebd26a8f6e2da3aad88a32ff73753926b3e41c3
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51410
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Edward Haas <edwardh@redhat.com>