Introduce host.py and commands.py to separate execCmd and AsyncProc from utils
execCmd and AsyncProc require cpopen and StringIO which are not available
in python3. utils.py should split to sub packages anyway, and this work
not only provides the ability to import utils.py with python3, but also
start the work of splitting utils.py to meaningful packages - host.py for
host's related tools, and commands for command execution tools.

Change-Id: I295f5516154c464fe5b1fc103b7b9d1d76dffe17
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51421
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Tested-by: Edward Haas <edwardh@redhat.com>