virt: Don't expose GuestAgent.guestInfo directly
GuestAgent.getGuestInfo() returns its guestInfo dict attribute directly,
so the callers may modify it.  guestInfo is a public attribute so it
shouldn't be a problem in theory as it's possible to modify that
attribute anyway without calling getGuestInfo().  But guestInfo isn't
used anywhere outside GuestAgent and should be probably changed to a
private attribute.

The problem of modifiable internal structure is exposed in
Vm._getGuestStats where modifications to guestInfo cause incorrect
memUsage values reported in case of repeated invocations (see the
referenced bug).

The safest approach to the problem is to return a deep copy of guestInfo
from getGuestInfo().  However, making a deep copy can have performance
issues, e.g. when copying a list of thousands installed Windows
packages.  For that reason we return just a shallow copy, which should
suffice for typical uses.

Change-Id: Ie48c697300e4feeb513dc662c2d1d3d8c21f0af9
Bug-Url: https://bugzilla.redhat.com/1295428
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51508
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-on: https://gerrit.ovirt.org/52335
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>