Refactor metadata operations
Volume metadata can be represented as a dictionary and as a string of
delineated key value pairs.  Extract this logic out of the Volume
classes into a new VolumeMetadata class which can be unit tested.  Once
refactored, this code can be reused by new SDM code as well.

Change-Id: I4fee56b30c13a3c1cede8489338ed60f4e1d5eab
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/52671
Continuous-Integration: Jenkins CI
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>