storage: Rename *VolumeMetadata to *VolumeManifest
When refactoring the storage classes (StorageDomain, Image, Volume) we
made the mistake of using 'Metadata' for the refactored volume classes
whereas we used 'Manifest' for image and the storage domains.  This
patch converts the *Volume classes to use Manifest as well.  The
VolumeMetadata class was well tested by sdm_indirection_tests.py and
after this patch the tests continue to pass.  Read on for more on the
rational for this change.

There are several reasons to make this change:

1. All of the other classes use Manifest and *Volume classes were the
   exception.  Code will be more readable with a common convention.

2. Originally 'Manifest' was chosen for *StorageDomain and Image because
   manifest means 'a list or invoice of goods' (such as you might
   receive in a delivered package.  StorageDomains have a list of images
   (also storage domain metadata and operations).  Images have a list of
   volumes (also operations).  Volumes do not contain a list of other
   objects (but they do have metadata, data, a lease, and operations).
   This list of components still carries a valid association with the
   concept of a manifest and therefore the name change is justified.

3. Volumes have their own metadata.  We want to factor out volume
   metadata manipulation into its own class.  What would the best name
   for such a factored out class be?  Why... VolumeMetadata!

Change-Id: I0a814b4cea34b49b2b7e72bb6ed25688ff05cce0
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51539
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Nir Soffer <nsoffer@redhat.com>