net: Normalize management network defaultRoute
The management network should always include defaultRoute=True.  With
ovirt-node<=3.5, the management network is initally set from the tui and
does not set defaultRoute.  defaultRoute is normalized internally to
True without adding it to the persistent config.

The fix introduces a normalization of the persistent config, such that
the management network will be marked with defaultRoute=True even if
defaultRoute is missing.

This issue is relevant for upgrades from ovirt-node<=3.5.

Change-Id: I3c596afaf6e57944962ebc874bd6a28374cfe382
Bug-Url: https://bugzilla.redhat.com/1293562
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51594
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>