net: Remove the defaultGateway api canonization
Previous versions (<3.6) have allowed the management network to be set
without defaultGateway=True from the API.  For the management network,
VDSM has canonized defaultGateway to True if the flag was missing.

From 3.6, defaultGateway is set explicitly by Engine and there is no
longer a need to canonize the value.

Change-Id: Ib6a062687fe39352fa49cbe7bd446e8d3cfa0c27
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51596
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI