Alignment scans will now run in direct mode
Currently on RHEL 7 and Fedora 22 or later alignment scan via libguestfs
does not work by going through libvirt and only works directly.  As a
workaround this patch runs the alignment scan with an environment
variable that tells libguestfs to run the scan directly and not through
libvirt.

While this solution works it should be reverted after the libvirt bug will
be solved as it is a workaround, since we allow only VMs in down status to
be scanned it does not have an affect but once we would like to allow scans
on running VMs the fact that we bypassing libvirt will mark the VMs as
tainted.

Bug-Url: https://bugzilla.redhat.com/1151838
Change-Id: I00d8b681e4d2a2f4fb60e673eefdc807d8e1447d
Signed-off-by: Tal Nisan <tnisan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40402
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Tested-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI