migration: use context manager for semaphore
The migration source code path needs to take the
ongoingMigrations semaphore, to throttle the outgoing
migrations.

The code did so using the pre-2.6 way, explicitly
using acquire() and release().

This patch switches the code to the context manager
semantics, to make the code a bit nicer and to unlock
further refactoring.

Change-Id: I8fb7bd6a66898804eefd9bd286c77fd8b9225bc4
Backport-To: 3.6
Backport-To: 3.5
Bug-Url: https://bugzilla.redhat.com/1298190
Related-To: https://bugzilla.redhat.com/1296936
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51655
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51827
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/52341