virt: safer handling of migration parameters
The Migration Source code path needs to store additional
transient migration parameters in the Vm configuration state,
to be both saved on recovery file and transmitted on the
destination VM.
Those parameters will be dropped once the migration is done.

Up until now, the migration code mangled the Vm state from the outside.
This was ugly and unsafe, because the configuration was mutated without
any locking.

Under high migration load, this can lead to RuntimeErrors, because
the clients of Vm.conf (e.g. status(), getVMList()...) expect the
lock being held and the access serialized.

To fix those issues, we add a helper method into Vm class, to
let the migration code temporarily store the data it needs in a safe
and simple way.

Change-Id: I063d74c1fe1725c514f7aeb60463d0558e240baa
Backport-To: 3.6
Backport-To: 3.5
Bug-Url: https://bugzilla.redhat.com/1296936
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51656
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51828
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>