net tests: Wait for bond dev state up
Several functional tests that involved bond devices, checked that
changes do not affect the bond state.

Tests that waited for an 'known' state, failed randomly, mainly when
the device state was down and during the test it came up.

The initial state for the bond state should be UP.
(this patch alligns all tests to use state UP as the pre-requirement)

Change-Id: I8da244d88c7c35bb41c72ab16e640e57cce80807
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51698
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>