imagetickets: Support image tickets
The introduced module is part of the image transfers implementation,
which lets the end user of oVirt to transfer images to and from
oVirt's storage devices.

Specifically, this VDSM module is responsible for granting permissions
for image transfers, via a ticket mechanism for an image, sent from
engine, through VDSM to imaged - a process for image IO which resides
next to VDSM in the host.

Change-Id: I6b9ded4bde73b1ab504cae50d2cea726d4f77e51
Signed-off-by: Amit Aviram <aaviram@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/50014
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI