virt: Add new guest agent API function for lifecycle-events
This patch introduces a new API method to the guest agent which allows
notifying the guest agent about selected lifecycle events.

The main purpose of this is to allow the guests to perform actions before
and after such events because a few selected events might have an impact
on them.

Change-Id: Idc2fac9daa065cb4f84289fe3910775813042096
Guest-Agent-API-Version: 3
Bug-Url: https://bugzilla.redhat.com/1298120
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51780
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI