vm: delete recovery file only after destroy()
We must not re-create a recovery file after it
is cancelled, because this will create an orphan
recovery file, and ghost VMs (see BZ below).

But we should also do not delete the recovery
file too early, because Engine is in charge
of collecting down VMs, and Vdsm is not really
done with VMs until that moment.

So, we move the cleanup of the recovery file
in deleteVm. We kill the recovery file only
when we wipe out the VM from the system, not earlier.

Change-Id: I9513a72041e9514fd3aed8b2292f2ebdea4a03ae
Bug-Url: https://bugzilla.redhat.com/1253043
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51799
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>