lun: Serial attr should not passed to libvirt for lun disks.
Following the change which was introduced at
libvirt-1.2.17-5 and qemu-kvm-rhev-2.3.0-18, libvirt now enforce
that a scsi-block 'lun' device should not provide the 'serial' property.

To comply with this change, this patch introduces a change
that a VM should ignore 'serial' property for lun devices when
passing the XML to libvirt.

This patch was verified by running a VM with lun disk which
had enabled SCSI Pass-Through and also plugging the same lun disk
while the VM was running. The operation succeeded.

Change-Id: I41abb2b7d7f5950e65710379a0655399014afd83
Bug-Url: https://bugzilla.redhat.com/1291930
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>
Tested-by: Maor Lipchuk <mlipchuk@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51766
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51848
Reviewed-by: Francesco Romani <fromani@redhat.com>