net: When IPv6 is not enabled, fallback to IPv4
There are setups where IPv6 is disabled, causing a conflict with the
default configuration (which is set to listen on address '::').
In such a case, VDSM fails to listen on '::'.

In order to support IPv4 only setups without requiring a change in the
configuration, VDSM will fallback to listen on an IPv4 only address.

Note that '::' is used to listen on both IPv4 & IPv6 stacks, a fallback to
IPv4 seems reasonable.

Change-Id: If95914dd9dfc88b40887345ac9e244ee2b1e09e9
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/52366
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>